fix: refactor selection methods to use population manager in algorithm
